Separate hit map toggle widget code from the typewriter panel.
